% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{inbase}
\alias{inbase}
\title{InBase inteins.}
\format{A tibble with 585 rows and 31 variables.}
\source{
\url{http://www.inteins.com}
}
\usage{
inbase
}
\description{
The InBase entries for all 585 inteins except for "Intein aa Sequence".
}
\keyword{datasets}
