% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{inbase_seq}
\alias{inbase_seq}
\title{Protein sequences of InBase inteins.}
\format{An AAStringSet with 585 rows in the same order as \code{inbase}.}
\source{
\url{http://www.inteins.com}
}
\usage{
inbase_seq
}
\description{
The "Intein aa Sequence" InBase entries of all 585 inteins.
}
\keyword{datasets}
